requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
langchain
langchain-core
langgraph
langsmith
langchain-openai
langchain-community
pandas
matplotlib
seaborn
pillow
gunicorn
rake-nltk
